refactor: Moved tests to their own directory
